+++ src/collector.rs
@@ -203,11 +203,11 @@ fn collect_and_save_data(
     if settings.need_to_refresh_processes {
         for process_opt in &process_cache_data.process_used {
             if let Some(process) = process_opt {
-                data_to_save.push(format!("{:.2}", process.cpu_usage));
+                data_to_save.push(format!("{:.2}", process.cpu_usage / 8.0));
                 data_to_save.push(convert_bytes_into_mega_bytes(process.memory_usage).to_string());
             } else {
-                data_to_save.push("0".to_string());
-                data_to_save.push("0".to_string());
+                data_to_save.push("-1".to_string());
+                data_to_save.push("-1".to_string());
             }
         }
     }
@@ -286,29 +286,36 @@ pub fn check_for_new_and_old_process_data(sys: &mut System, process_cache_data: 
 }
 
 fn check_which_process_to_track(process_cache_data: &mut ProcessCache, sys: &mut System, settings: &Settings, system_pids: &HashSet<usize>) {
-    for (pid, process) in sys.processes() {
-        let pid_number: usize = (*pid).into();
-        if process_cache_data.processes_checked_to_be_used.contains(&pid_number) || !system_pids.contains(&pid_number) {
+    for (idx, i) in settings.process_cmd_to_search.iter().enumerate() {
+        if process_cache_data.process_used[idx].is_some() {
+            // Already monitoring process from such name
             continue;
         }
 
-        let collected_name = process.cmd().join(" ");
-        for (idx, i) in settings.process_cmd_to_search.iter().enumerate() {
-            if process_cache_data.process_used[idx].is_some() {
-                // Already monitoring process from such name
+        let mut shortest_matching_process_data = None;
+        let mut shortest_text = usize::MAX;
+
+        for (pid, process) in sys.processes() {
+            let pid_number: usize = (*pid).into();
+            if process_cache_data.processes_checked_to_be_used.contains(&pid_number) || !system_pids.contains(&pid_number) {
                 continue;
             }
-            if collected_name.contains(&i.search_text) {
-                info!(
-                    "Found process \"{}\" with pid \"{}\" that will be monitored - (\"{}\")",
-                    process.name(),
-                    pid_number,
-                    collected_name,
-                );
-                process_cache_data.processes_checked_to_be_used.insert(pid_number);
-                process_cache_data.process_used[idx] = Some(CustomProcessData::from_process(process));
-                break;
+            let collected_name = process.cmd().join(" ");
+            if collected_name.contains(&i.search_text) && collected_name.len() < shortest_text {
+                shortest_text = collected_name.len();
+                shortest_matching_process_data = Some((pid_number, process, collected_name));
             }
+        }
+
+        if let Some((pid_number, process, collected_name)) = shortest_matching_process_data {
+            info!(
+                "Found process \"{}\" with pid \"{}\" that will be monitored - (\"{}\")",
+                process.name(),
+                pid_number,
+                collected_name,
+            );
+            process_cache_data.processes_checked_to_be_used.insert(pid_number);
+            process_cache_data.process_used[idx] = Some(CustomProcessData::from_process(process));
         }
     }
 }
@@ -339,6 +346,7 @@ fn remove_tracking_of_removed_processes(process_cache_data: &mut ProcessCache, s
 // Needed to get processes name and cmd, rest is updated in update_usage_of_tracked_process
 fn update_new_processes_stats(process_cache_data: &mut ProcessCache, sys: &mut System, system_pids: &HashSet<usize>) {
     let new_processes = process_cache_data.get_differences_in_usage_processes(system_pids.iter());
+
     if !new_processes.is_empty() {
         if new_processes.len() > 40 {
             info!("Found {} new processes, refreshing them in batch", new_processes.len());
@@ -354,6 +362,10 @@ fn update_new_processes_stats(process_cache_data: &mut ProcessCache, sys: &mut S
 }
 
 fn update_usage_of_tracked_process(process_cache_data: &mut ProcessCache, sys: &mut System) {
+    let process_count = process_cache_data.process_used.iter().flatten().count();
+    if process_count == 0 {
+        return;
+    }
     debug!("Updating data of {} processes", process_cache_data.process_used.iter().flatten().count());
     for custom_process in process_cache_data.process_used.iter_mut().flatten() {
         sys.refresh_process_specifics(Pid::from(custom_process.pid), ProcessRefreshKind::new().with_cpu());